did  lot of changes and tested all end points

+++ src/main/java/comproject/jobApp/Repo/JobRepo.java
@@ -4,6 +4,10 @@ import comproject.jobApp.model.Job;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+import java.util.OptionalLong;
+
 @Repository
-public interface JobRepo extends JpaRepository<Job, Integer> {
+public interface JobRepo extends JpaRepository<Job, Long> {
+
 }
